+++ lib/netsuite/records/item_vendor.rb
@@ -2,10 +2,19 @@ module NetSuite
   module Records
     class ItemVendor
       include Support::Fields
+      include Support::RecordRefs
       include Support::Records
       include Namespaces::ListAcct
 
-      fields :vendor, :purchase_price, :preferred_vendor
+      fields :purchase_price,
+            :preferred_vendor,
+            :vendor_code,
+            :vendor_currency_name
+
+      record_refs :schedule,
+                  :subsidiary,
+                  :vendor,
+                  :vendor_currency
 
       def initialize(attributes = {})
         initialize_from_attributes_hash(attributes)